When client says "everyone is a citizen", also update on members
Previously in this step on the SNAP flow, when the client selected "everyone is a citizen", we only marked that on the application and didn't update individual members. As a result, when we fill in the PDF, the members were never marked as citizens on the PDF, because we only fill it out based on the member.

This PR fixes that by updating members-> citizen to true when 'everyone_a_citizen' is selected.

This PR also addresses the negative of that problem on the same step, which happens when someone marks 'anyone_blank' as false. Right now we aren't updating the members (as a result, the value for that field is nil). This PR marks those fields as false instead.

[#154776076]