Refactor 'member_attributes' and 'application_attributes'
Creates new 'form' class (based off of 'step') that uses a combination of simple 'member_attributes' and 'application_attributes' to construct one set of common form attributes. Getters for these methods can then be used to filter params as appropriate for their respective models.

For now, 'member_attributes' and 'application_attributes' only accept keys or strings as attributes, unlike 'step_attributes' method, which accepts hashes. We can add this in later as we need it.

[#155522274]