Make presence validation error messages clearer
In doing the previous commit, I noticed that our error messages were a little hard to parse since the field name wasn't included in the message, e.g. not selecting a radio button just showed the error message "can't be blank". Changed this to be a bit more verbose to bring them in line with other error messages, and to suggest a remedy for a failed presence validation that may not be because of user error.

[#156055105]