Reduce number of duplicative integration tests
We're unit testing our PDF generation and attribute filling, so testing via integration test is a bit duplicative. This leverages integration tests only to verify the PDF has been sent and is not corrupted (a problem we've had in the past). Also, it does this only in one test, the most complicated with each supplement included. We may want to move the PDF assembly tests into a unit test for the PdfAssembler and only rely on the integration test to check for lack of corruption, but this is a start.

[#155715243]