Ask applicants which office they're in for SNAP + Medicaid flows
We were receiving misdirected applications, because offices aren't always using the office-specific landing places. To prevent misdirected applications, we added a question that explicitly asks users where they're applying. If they indicate that they're not in an office or are in an office other than those we serve, we direct the application to us. Otherwise, we route to the correct office.

To do this, we now have a synthetic "office_location" field on the old-style applications, which looks first to a recognized chosen office location, then to whether someone applied via an office page. If neither of these results in an office, then we look up the person's address by county to route it.

[Finishes #156967567]

Co-authored-by: Ben Golder <bgolder@codeforamerica.org>