Use 2FA-specific env var to determine 2FA environment
Instead of relying on RAILS_ENV, which can be the same across multiple deployment environments. Each deployment environment needs a unique name so that we can create a unique 2FA name and code.

This divorces RAILS_ENV from that 2FA slug, and will allow us to have consistent RAILS_ENV across staging, demo, and production, which will come in a later commit.

Does not modify the naming of any existing 2FA environment titles, so as not to disrupt any existing 2FA codes.

[#158276758]