Add 2fa demo env (#829)
* Use 2FA-specific env var to determine 2FA environment

Instead of relying on RAILS_ENV, which can be the same across multiple deployment environments. Each deployment environment needs a unique name so that we can create a unique 2FA name and code.

This divorces RAILS_ENV from that 2FA slug, and will allow us to have consistent RAILS_ENV across staging, demo, and production, which will come in a later commit.

Does not modify the naming of any existing 2FA environment titles, so as not to disrupt any existing 2FA codes.

[#158276758]

* Remove reference to RAILS_ENV staging env

We no longer need a staging RAILS_ENV, as we were only relying on it to set our 2FA name. This now uses an independent variable, so we should change our RAILS_ENV to all be production when deployed to make them as similar as possible.

[Finishes #158276758]

* Don't prompt to configure 2FA in development

[#158276758]

* Remove remnant staging environment config

Every deployed environment now uses 'production' env.

[#158276758]