Reduce default wait time for Capybara to 0
With the upgrade from Capybara 2 to 3, the behavior for `first` and `all` changed, causing our axe-matchers gem to cause tests with accesibility checks to run 7x slower or more. To fix this, we set the default wait for Capybara to 0, since we're not actually waiting on async responses.

This also begins using capybara-webkit as our driver for axe-matchers/js, since it's a bit faster.

[#158733145]