Only run accessibility specs on CI
We've reduced the total spec suite time by reducing Capybara's default time to wait, but it still takes 11 min to run. The tests that take the longest to run are our feature specs, which are slowed down by the accessibility checks. To keep our local spec runs relatively speedy, this allows users (or CI) to opt into running the a11y checks.

[#158733145]