Update CircleCI config to 2.0 format
Also parallelizes our jobs by taking advantage of the new workflows feature. On Circle, we should now have:

On non-master branches:
* a build job
* parallelized: bundle audit & rubocop, rspec

On master branch:
* a build job
* rspec
* if above passes, deploy to staging

We should be able to retrigger any one of those tasks individually if they fail.

This also adds a dependency on a Docker image with pre-installed dependencies, which we'll need to update at least whenever we update Ruby versions. CONTRIBUTING.md has been updated with instructions to do so.