Show selected radio button when there are errors
Fixes bug in our JS that didn't show selected state for radio buttons when errors were shown, because of our HTML structure. JQuery now looks for an ancestor with '.radio-button' instead of just a parent.

Also, to get selected state to show with our Javascript-injected partial, re-attach the event handlers to the new HTML.

[Finishes #159703918]